Move Integration into android package.
