+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Move `FormulaFragment` into `android` package.
 - Move `BackCallback` into `android` package.
 - Move `FragmentLifecycleEvent` into `android.events` package.
+- Move `Integration` into `android` package.
 
 ## [0.6.1] - November 18, 2020
 - Bugfix: Fix runtime ignoring `Formula.key` for the root formula.
